#56 default emptied
