+++ examples/atmega328p/analog/src/main.rs
@@ -3,9 +3,9 @@
 #![deny(warnings)]
 
 /// Crates to be used.
-use rustduino::atmega328p::hal::analog::{map_from1023_to255, RefType};
-use rustduino::atmega328p::hal::pins::Pins;
-use rustduino::atmega328p::hal::watchdog::Watchdog;
+use rustduino::hal::analog::map_from1023_to255;
+use rustduino::hal::pin::Pins;
+use rustduino::hal::watchdog::Watchdog;
 
 #[no_mangle]
 pub fn main() {
@@ -19,7 +19,7 @@ pub fn main() {
     // Infinite loop for read and write continuously through the I/O pins.
     loop {
         // Take input into the zeroth analog pin.
-        let a: u32 = pins.analog[0].read(RefType::DEFAULT);
+        let a: u32 = pins.analog[0].read();
 
         // Make the input value ready to be sent through a digital pin.
         let b: u8 = map_from1023_to255(a);
