+++ src/atmega2560p/hal/port.rs
@@ -19,7 +19,7 @@
 //! https://ww1.microchip.com/downloads/en/devicedoc/atmel-2549-8-bit-avr-microcontroller-atmega640-1280-1281-2560-2561_datasheet.pdf
 
 /// Source codes required.
-use crate::atmega2560p::hal::pin::{AnalogPin, DigitalPin};
+use crate::hal::pin::{AnalogPin, DigitalPin};
 
 /// Core Crate functions required in the code for reading and writing to registers.
 use core::{
