+++ src/atmega2560p/hal/watchdog.rs
@@ -19,7 +19,7 @@
 //! https://ww1.microchip.com/downloads/en/devicedoc/atmel-2549-8-bit-avr-microcontroller-atmega640-1280-1281-2560-2561_datasheet.pdf
 
 /// Crates required in the code for reading and writing to registers.
-use crate::atmega2560p::hal::interrupts;
+use crate::hal::interrupts;
 use core::ptr::{read_volatile, write_volatile};
 
 /// Use interrupts to enable/disable global interrupts,
